Move the vendor details templates to the index route
This makes things easier to manage and we only have a single nested route so it doesn't affect anything else.
