+++ app/controllers/vendors/details/index.js
@@ -7,14 +7,65 @@ import { task } from 'ember-concurrency';
 export default class VendorsSubjectController extends Controller {
   @service router;
   @service store;
-  @tracked bestuurseenheidToRemove;
   @tracked vendor;
+  @tracked bestuurseenhedenLijst = [];
+  @tracked isAddingAdministrativeUnits = false;
+  @tracked bestuurseenheidToRemove;
   sort = 'naam';
   page = 0;
   size = 20;
 
   get shouldShowDeleteConfirmationModal() {
     return Boolean(this.bestuurseenheidToRemove);
+  }
+
+  removeFromList = task(async () => {
+    const bestuurseenheid = this.bestuurseenheidToRemove;
+    const vendors = await bestuurseenheid.vendors;
+    vendors.splice(vendors.indexOf(this.vendor), 1);
+    await bestuurseenheid.save();
+    //We must trigger model(), since the pagination depends on this.
+    this.router.refresh('vendors.details.index');
+    this.hideDeleteConfirmationModal(true);
+  });
+
+  addToList = task({ drop: true }, async () => {
+    await Promise.all(
+      this.bestuurseenhedenLijst.map(async (bestuurseenheid) => {
+        const vendors = await bestuurseenheid.vendors;
+        vendors.push(this.vendor);
+        await bestuurseenheid.save();
+      })
+    );
+
+    this.bestuurseenhedenLijst = [];
+    this.router.refresh('vendors.details');
+    this.closeAddModal();
+  });
+
+  @action
+  async appendBestuurseenheid(eenheid) {
+    this.bestuurseenhedenLijst.pushObject(eenheid);
+  }
+
+  @action
+  removeBestuurseenheid(eenheid) {
+    this.bestuurseenhedenLijst = this.bestuurseenhedenLijst.without(eenheid);
+  }
+
+  @action
+  copyToClipboard(key) {
+    navigator.clipboard.writeText(key);
+  }
+
+  @action
+  showAddModal() {
+    this.isAddingAdministrativeUnits = true;
+  }
+
+  @action
+  closeAddModal() {
+    this.isAddingAdministrativeUnits = false;
   }
 
   @action
@@ -28,14 +79,4 @@ export default class VendorsSubjectController extends Controller {
       this.bestuurseenheidToRemove = null;
     }
   }
-
-  removeFromList = task(async () => {
-    const bestuurseenheid = this.bestuurseenheidToRemove;
-    const vendors = await bestuurseenheid.vendors;
-    vendors.splice(vendors.indexOf(this.vendor), 1);
-    await bestuurseenheid.save();
-    //We must trigger model(), since the pagination depends on this.
-    this.router.refresh('vendors.details.index');
-    this.hideDeleteConfirmationModal(true);
-  });
 }
