--- app/templates/vendors/details.hbs
@@ -1,53 +1,0 @@
-<div class="au-o-body-container au-o-box au-o-layout au-u-margin-top au-u-margin-bottom">
-    <AuHeading class="au-o-region">{{t 'vendor.subject.main.head-1'}}</AuHeading>
-
-    <AuCard @flex="true" class="au-u-margin-bottom" as |c|>
-      <c.header @badgeSkin="brand" @badgeIcon="users">
-        <AuHeading @level="2" @skin="4">
-          {{this.model.name}} ({{this.model.uri}})
-        </AuHeading>
-
-
-      </c.header>
-      <c.content>
-        <div class="au-o-box">
-          <AuLabel for="input-regular" class="au-u-bold">{{t 'vendor.subject.card.head-1'}}</AuLabel>
-          <AuInput id="input-regular" @type={{if this.hiddenKeyState "text" "password"}}
-            @value="{{this.model.key}}" readonly />
-          <AuButton {{on 'click' (toggle "hiddenKeyState" this)}}>{{t 'vendor.subject.card.button-1'}}</AuButton>
-          <AuButton {{on 'click' (fn this.copyToClipboard this.model.key)}} @skin="secondary">{{t 'vendor.subject.card.button-2'}}</AuButton>
-        </div>
-      </c.content>
-    </AuCard>
-    <AuToolbar class="au-o-region" as |Group|>
-      <Group>
-
-        <AuHeading @skin='3'>{{t 'vendor.subject.table.head-1'}}</AuHeading>
-
-      </Group>
-      <Group>
-        <AuButton @icon="add" {{on "click" this.showAddModal}}>{{t "vendor.subject.table.button-1"}}</AuButton>
-      </Group>
-    </AuToolbar>
-        {{outlet}}
-</div>
-
-{{#if this.isAddingAdministrativeUnits}}
-  <AuModal @modalOpen={{true}} @closeModal={{this.closeAddModal}} @overflow={{true}}>
-    <:title>
-      {{t 'vendor.subject.modal.head-1'}}
-    </:title>
-    <:body>
-      <BestuurseenheidToevoegen
-        @onListBestuurseenheden={{this.listBestuurseenheden}}
-        @appendBestuurseenheid={{this.appendBestuurseenheid}}
-        @bestuurseenhedenLijst={{this.bestuurseenhedenLijst}}
-        @removeBestuurseenheid={{this.removeBestuurseenheid}} />
-    </:body>
-    <:footer>
-      <AuButton @loading={{this.addToList.isRunning}} @loadingMessage="Aan het toevoegen" {{ on 'click' (perform this.addToList)}}>
-        {{t 'vendor.subject.modal.button-1'}}
-      </AuButton>
-    </:footer>
-  </AuModal>
-{{/if}}
